📝 Prepare docs for new release (#156)
## Description

This PR just fixes up a couple of things throughout the documentation
regarding the move from `qecc` to `mqt-qecc` and from the `QFR` library
to `MQT Core`. It also slightly updates the RtD documentation to bring
it more in line with QCEC and QMAP.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.