Ftsp (#234)
## Description

This PR adds methods for synthesizing and simulating fault-tolerant
state preparation circuits for CSS codes.

The main functionality introduced is:

- Optimal and heuristic synthesis of non-ft state preparation circuits
- Optimal and heuristic verification of a non-ft state preparation
circuit
- Simulation methods to numerically estimate the logical error rate of a
state-preparation circuit

Furthermore, this PR introduces a `CSSCode` class for convenience and
adapts existing code implementations to conform to this class. For
example, the `HexagonalColorCode` class now inherits from `CSSCode`.
Some sample CSS codes are provided that can be directly imported via the
`CSSCode` class.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.