Don't bound bitvector values if n is a power of 2.
If a code has a number of qubits that is a power of 2, then the assertion in `_generate_circ_with_bounded_gates` that
makes sure the bitvector variables encoding the CNOT qubits act only on valid qubits erroneously asserts that control
and target should be less than 0. This of course always returns `False`.