Differentiate error correction and error detection circuits.
So far we assumed that we always want to have an error correction code and thus the maximum number of independent errors
in the circuit was computed as (d-1)//2. But for an error detection code we allow d//2 independent errors.

For this I introduced a flag `error_detection_code` in the `StatePrepCircuit` class so that the errors are computed
correctly.

Furthermore, if d=2 this flag is set automatically.