Fix small code bugs for FT Stateprep (#267)
## Description

This PR fixes some issues with certain codes. 

There was a bug when the code was defined over a number of qubits, which
is an exact power of 2, leading to the optimal state prep circuit
failing to be synthesized.

Another issue occurred when looking at d=2 codes, which cannot be used
for error correction but only for error detection. I have added
functionality to set whether one wants to create the circuit for error
detection or error correction in the `StatePrepCircuit` class.

Fixes #265

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.