Fix error propagation. (#283)
## Description

There is an error in the computation of fault sets in state preparation
circuits. Sometimes, the gates are not traversed in the correct
topological order, and thus, some faults are not correct. This didn't
show up before because the faults were still caught by the verification
circuits.

Fixes # (issue)

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.