Adjustments to handling of hook errors in FT state preparation circuit synthesis (#285)
## Description

Currently, there is a bug in the handling of hook errors where they are
not computed quite correctly. Unfortunately, this didn't come up in
previous simulations and was therefore missed.

In the process of fixing this and resynthesizing the current circuits, I
realized that sometimes, flagging the first verification layer leads to
better overall results than handling hook errors in the second
verification layer. Therefore, this PR introduces an option in the state
preparation synthesis that chooses whether the first layer should be
hooked or not.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: lsschmid <117631861+lsschmid@users.noreply.github.com>