Simulation steane ftsp (#383)
## Description

This PR introduces a simulation class to simulate Steane-type
fault-tolerant state preparation circuits using four circuits.

Since this simulation method has a lot of overlapping functionality with
our previous simulation setup, this PR also refactors the common
functionality into the abstract class `NoisyNDFTStatePrepSimulator`. The
implementation of the old simulator is now in
`VerificationNDFTStatePrepSimulator`.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [ ] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>