reference based heuristic changed to reference fault sets and speed ups
Now we can construct a reference circuit based on fault sets that we
provide. Before only the fault set of the given reference SPC was used.
It is now possible to also find circuits based on other fault sets.

We have introduced some speed ups to the function. We save now errors
that have shown to be overlapping to prevent recalculation of those
errors. We also now only check the overlap with the new error and not
with the whole provisional fault set anymore.