Codes from file (#389)
## Description

This PR introduces functionality to load a `CSSCode` or `StabilizerCode`
instance from a file that lists the stabilizer generators as strings
with one stabilizer generator per line.

Furthermore, this PR removes the necessary distance argument from the
`CSSCode` class. Since `d` was the first argument in the constructor
before, some files and the documentation needed to be refactored.


## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.