Add new logo banner w/o MQT logo and restructure main documents (#411)
## Description

This PR takes the logo banner and removes the MQT logo. It adds the
resulting banner as new files for light and dark mode.

Additionally, it refines the wording around MQSC and restructures the
README as well as the index page.


## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.