Ignore qsample warnings. (#426)
## Description

Recently, qsample has been made an optional dependency, which means that
tests relying on qsample are skipped in the CI and don't lead to
failures due to DeprecationWarnings anymore. However, if qsample is
installed locally and tests are run, the tests will still fail.

This PR fixes this issue by ignoring the warnings thrown by qsample.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [ ] I have made sure that all CI jobs on GitHub pass.
- [ ] The pull request introduces no new warnings and follows the
project's style guidelines.